+++ example/app/web/flutter_service_worker.js
@@ -4,9 +4,9 @@ const TEMP = 'flutter-temp-cache';
 const CACHE_NAME = 'flutter-app-cache';
 const RESOURCES = {
   "version.json": "11e1a177d57c5da2bc6e3a6acc3010f4",
-"index.html": "ebea9d3b7f82f1ef612963faf5eebce0",
-"/": "ebea9d3b7f82f1ef612963faf5eebce0",
-"main.dart.js": "67ef1e726c4b33ae1340e22f500320cd",
+"index.html": "d7f1a567af1a6066f9bd21d7ca8f5db6",
+"/": "d7f1a567af1a6066f9bd21d7ca8f5db6",
+"main.dart.js": "dd5e7a7b1f968ec12be092b8f0b90ee0",
 "favicon.png": "7b1e138b59ab213ab88128833a2681f3",
 "icons/Icon-192.png": "786a784711f7ba2dcc1aa8f4fa52d02b",
 "icons/Icon-512.png": "8a74e06ad551887241db3b8c87b78697",